refactor: extract CliProcessor
